+++ .idea/modules.xml
@@ -3,7 +3,9 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/jibcon/jibcon.iml" filepath="$PROJECT_DIR$/jibcon/jibcon.iml" />
+      <module fileurl="file://$PROJECT_DIR$/jibcon/jibcon.iml" filepath="$PROJECT_DIR$/jibcon/jibcon.iml" />
       <module fileurl="file://$PROJECT_DIR$/jibcon_android.iml" filepath="$PROJECT_DIR$/jibcon_android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/jibcon_android_0529.iml" filepath="$PROJECT_DIR$/jibcon_android_0529.iml" />
     </modules>
   </component>
 </project>
