+++ jibcon/src/main/res/layout/data_control.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
+    android:layout_gravity="center"
+    android:background="#FFF">
+
+
+    <ListView
+        android:id="@+id/Lv_data_control"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+</LinearLayout>
